renaming to wcc + refactor numa node out of module tree to remove libnuma dependency
